+++ tests/test_function.py
@@ -1,4 +1,5 @@
 """Test kedro_kubeflow functions."""
 
+
 def test_app():
     pass