+++ CHANGELOG.md
@@ -2,16 +2,24 @@
 
 ## [Unreleased]
 
+## [0.1.9] - 2021-01-08
+
+-   Support for MLFlow - if the package is installed then additional step is added with parent run init. Then all separate nodes runs register under this run.
+-   Support for inter-steps volume: setup (one volume per pipeline run), initial load (the content of `data/` directory from the image and mount to all the steps for artifacts passing.
+-   `kubeflow init` command added to generate sample configuration file.
+
 ## [0.1.8] - 2021-01-05
 
 ### Added
 
--   *Initial release* of kedro-kubeflow plugin
+-   _Initial release_ of kedro-kubeflow plugin
 -   Ability to run an anonymous pipeline once as within a specified experiment `kedro kubeflow run-once`.
--   Ability to upload pipeline `kedro kubeflow upload`.
+-   Ability to upload pipeline `kedro kubeflow upload-pipeline`.
 -   Method to schedule runs for most recent version of given pipeline `kedro kubeflow schedule` 
 -   Shortcut to open UI for pipelines using `kedro kubeflow ui` 
 
-[Unreleased]: https://github.com/getindata/kedro-kubeflow/compare/0.1.8...HEAD
+[Unreleased]: https://github.com/getindata/kedro-kubeflow/compare/0.1.9...HEAD
+
+[0.1.9]: https://github.com/getindata/kedro-kubeflow/compare/0.1.8...0.1.9
 
 [0.1.8]: https://github.com/getindata/kedro-kubeflow/compare/ea219ae5f70e726b7afc9d0864da4b6649e4deaf...0.1.8